move server class
