+++ pom.xml
@@ -140,7 +140,7 @@
               <transformers>
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
                 <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                  <mainClass>com.mercari.solution.api.mcp.PipelineMcpStreamableServer</mainClass>
+                  <mainClass>com.mercari.solution.server.PipelineMcpStreamableServer</mainClass>
                 </transformer>
               </transformers>
             </configuration>
