+++ src/main/java/com/mercari/solution/server/mcp/tool/ValidatePipelineTool.java
@@ -1,16 +1,16 @@
-package com.mercari.solution.api.mcp.tool;
+package com.mercari.solution.server.mcp.tool;
 
-import com.mercari.solution.api.PipelineService;
+import com.mercari.solution.server.api.PipelineService;
 import io.modelcontextprotocol.server.McpSyncServerExchange;
 import io.modelcontextprotocol.spec.McpSchema;
 import jakarta.servlet.ServletContext;
 
 
 @Tool.Module(
-    name="run-pipeline",
-    title="Run Pipeline",
+    name="validate-pipeline",
+    title="Validate Pipeline",
     description= """
-        Run the pipeline defined in config parameter.
+        Validate the pipeline defined in config parameter.
         Results are returned in JSON format.
         If the pipeline definition is valid, outputs the output schema for each step in 'outputs' attribute.
         If the pipeline definition has errors, the error content is output in 'errors' attribute.
@@ -28,22 +28,9 @@ import jakarta.servlet.ServletContext;
         }
         """,
     outputSchema = """
-        {
-          "type": "object",
-          "properties": {
-            "status": {
-              "type": "string",
-              "description": "Definition of pipeline. YAML or JSON format."
-            },
-            "outputs": {
-              "type": "object"
-            }
-          },
-          "required": ["config"]
-        }
         """
 )
-public class RunPipelineTool implements Tool {
+public class ValidatePipelineTool implements Tool {
 
     @Override
     public void init(ServletContext servletContext) {
@@ -56,11 +43,11 @@ public class RunPipelineTool implements Tool {
             final McpSchema.CallToolRequest request) {
 
         if(!request.arguments().containsKey("config")) {
-            return new McpSchema.CallToolResult("run-pipeline mcp tool requires config parameter", true);
+            return new McpSchema.CallToolResult("validate-pipeline mcp tool requires config parameter", true);
         }
 
         final String config = request.arguments().get("config").toString();
-        final PipelineService.RunResult result = PipelineService.run(config, null, false);
+        final PipelineService.RunResult result = PipelineService.run(config, null, true);
         return McpSchema.CallToolResult.builder()
                 .addTextContent(result.responseText)
                 .isError(result.isError)
