+++ src/main/webapp/WEB-INF/web.xml
@@ -7,29 +7,29 @@
     <display-name>Mercari Pipeline Server</display-name>
 
     <servlet>
-        <servlet-name>MPipelineServer</servlet-name>
-        <servlet-class>com.mercari.solution.MPipelineServer</servlet-class>
+        <servlet-name>MPipelineApiServer</servlet-name>
+        <servlet-class>com.mercari.solution.server.PipelineApiServer</servlet-class>
         <async-supported>true</async-supported>
     </servlet>
 
     <servlet>
         <servlet-name>MPipelineMcpServer</servlet-name>
-        <servlet-class>com.mercari.solution.api.mcp.PipelineMcpStreamableServer</servlet-class>
+        <servlet-class>com.mercari.solution.server.PipelineMcpStreamableServer</servlet-class>
         <async-supported>true</async-supported>
     </servlet>
 
     <servlet-mapping>
-        <servlet-name>MPipelineServer</servlet-name>
+        <servlet-name>MPipelineApiServer</servlet-name>
         <url-pattern>/api/pipeline/*</url-pattern>
     </servlet-mapping>
 
     <servlet-mapping>
-        <servlet-name>MPipelineServer</servlet-name>
+        <servlet-name>MPipelineApiServer</servlet-name>
         <url-pattern>/api/schema/*</url-pattern>
     </servlet-mapping>
 
     <servlet-mapping>
-        <servlet-name>MPipelineServer</servlet-name>
+        <servlet-name>MPipelineApiServer</servlet-name>
         <url-pattern>/probe/*</url-pattern>
     </servlet-mapping>
